--- src/main/java/com/mikhailkarpov/vidly/vidlyapi/exception/AccountAlreadyExistsException.java
@@ -1,12 +1,0 @@
-package com.mikhailkarpov.vidly.vidlyapi.exception;
-
-public class AccountAlreadyExistsException extends RuntimeException {
-
-    public AccountAlreadyExistsException(String message) {
-        super(message);
-    }
-
-    public AccountAlreadyExistsException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
