+++ src/main/java/com/mikhailkarpov/vidly/vidlyapi/exception/UserAlreadyExistsException.java
@@ -1,0 +1,12 @@
+package com.mikhailkarpov.vidly.vidlyapi.exception;
+
+public class UserAlreadyExistsException extends RuntimeException {
+
+    public UserAlreadyExistsException(String message) {
+        super(message);
+    }
+
+    public UserAlreadyExistsException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
