+++ src/main/java/com/mikhailkarpov/vidly/vidlyapi/web/controller/RegistrationController.java
@@ -1,0 +1,32 @@
+package com.mikhailkarpov.vidly.vidlyapi.web.controller;
+
+import com.mikhailkarpov.vidly.vidlyapi.exception.UserAlreadyExistsException;
+import com.mikhailkarpov.vidly.vidlyapi.service.UserService;
+import com.mikhailkarpov.vidly.vidlyapi.web.dto.AccountDetails;
+import com.mikhailkarpov.vidly.vidlyapi.web.dto.AccountRegistrationRequest;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("api/accounts")
+public class RegistrationController {
+
+    private UserService userService;
+
+    public RegistrationController(UserService userService) {
+        this.userService = userService;
+    }
+
+    @PostMapping
+    public ResponseEntity<Object> register(@RequestBody AccountRegistrationRequest request) {
+        try {
+            AccountDetails account = userService.register(request);
+            return ResponseEntity.ok().body(account);
+        } catch (UserAlreadyExistsException e) {
+            return ResponseEntity.badRequest().body(e.getMessage());
+        }
+    }
+}
